[BR-5041] Fixed warnings in One2MBeanRelationshipTest on Oracle
The teardown phase of the test did or did not produce warnings depending on the currently used database. The warnings were not detected by PHPUnit since they were not triggered as part of the test body.

+++ sugarcrm/tests/{old}/SugarTestKBContentUtilities.php
@@ -53,16 +53,29 @@ class SugarTestKBContentUtilities
     {
         $db = DBManagerFactory::getInstance();
 
-        $db->query('DELETE FROM kbdocuments WHERE id IN (' . static::getPreparedIdsString(self::$_createdKbDocumentIds) . ')');
-        $db->query('DELETE FROM kbarticles WHERE id IN (' . static::getPreparedIdsString(self::$_createdKbArticleIds) . ')');
+        if (self::$_createdKbDocumentIds) {
+            $db->query('DELETE FROM kbdocuments WHERE id IN ('
+                . static::getPreparedIdsString(self::$_createdKbDocumentIds)
+                . ')');
 
-        $conditions = static::getPreparedIdsString(self::$_createdKbContentIds);
-        $db->query('DELETE FROM kbcontents WHERE id IN (' . $conditions . ')');
-        $db->query('DELETE FROM kbcontents_audit WHERE parent_id IN (' . $conditions . ')');
+            self::$_createdKbDocumentIds = array();
+        }
 
-        self::$_createdKbDocumentIds = array();
-        self::$_createdKbArticleIds = array();
-        self::$_createdKbContentIds = array();
+        if (self::$_createdKbArticleIds) {
+            $db->query('DELETE FROM kbarticles WHERE id IN ('
+                . static::getPreparedIdsString(self::$_createdKbArticleIds)
+                . ')');
+
+            self::$_createdKbArticleIds = array();
+        }
+
+        if (self::$_createdKbContentIds) {
+            $conditions = static::getPreparedIdsString(self::$_createdKbContentIds);
+            $db->query('DELETE FROM kbcontents WHERE id IN (' . $conditions . ')');
+            $db->query('DELETE FROM kbcontents_audit WHERE parent_id IN (' . $conditions . ')');
+
+            self::$_createdKbContentIds = array();
+        }
     }
 
     /**
